+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Metrics
 
 ### General
+- updated docs wrt. usage of `eval_framework.utils.generate_task_docs`
+- Added wandb logging and registry checkpoint loading with huggingface and vllm backends
 - Removed `torch` as a main dependency of `eval_framework`
 - Updated docs wrt. usage of `eval_framework.utils.generate_task_docs`
 
@@ -26,6 +28,10 @@
 - Added `subjects`and `hf_revision` to BaseTask arguments to replace global task re-definition when running with non default values.
 - Generate task documentation in `docs/tasks`. Moves the generate_task_docs utility to inside the package and added test that documentation is up-to-date.
 - Renamed `ChemBenchMultipleChoice` to `ChemBench` for consistency.
+- We've removed all models except those used for testing (they were largely old). The recommended way going forward is to provide
+  your own models implementation to the framework.
+- Removed and relaxes several main-dependencies
+- Moves the generate_task_docs utility to inside the package and add test that documentation is up-to-date
 - Fixed `ZERO_SCROLLS_QMSUM` missing from task_names.py
 - Fix inconsistent language code for Croatian/Serbian in INCLUDE task